Adds cookie policy page and UI
Implements a cookie policy page with detailed information about the cookies used on the website.

Introduces an accordion component for better organization and readability of cookie information.

Updates the icon library with new icons used in the cookie policy page.

Adds styling for the cookie policy page.

Updates footer menu with link to the cookie policy.